GameEvents: Move Game events to their own namespace
