+++ Snowflake.Events/CoreEvents/Game/GameProcessQuitEventArgs.cs
@@ -8,7 +8,7 @@ using Snowflake.Service;
 using Snowflake.Emulator;
 using Snowflake.Game;
 
-namespace Snowflake.Events.CoreEvents
+namespace Snowflake.Events.CoreEvents.GameEvent
 {
     public class GameProcessQuitEventArgs : GameEventArgs
     {
