+++ Snowflake.Events/CoreEvents/Game/GameQuitEventArgs.cs
@@ -3,24 +3,21 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using System.Diagnostics;
 using Snowflake.Service;
 using Snowflake.Emulator;
 using Snowflake.Game;
 
-namespace Snowflake.Events.CoreEvents
+namespace Snowflake.Events.CoreEvents.GameEvent
 {
-    public class GameProcessStartEventArgs : GameEventArgs
+    public class GameQuitEventArgs : GameEventArgs
     {
         public IEmulatorAssembly GameEmulatorAssembly { get; private set; }
         public IEmulatorBridge GameEmulatorBridge { get; private set; }
-        public Process GameEmulatorProcess { get; private set; }
-        public GameProcessStartEventArgs(ICoreService eventCoreInstance, IGameInfo gameInfo, IEmulatorAssembly emulatorAssembly, IEmulatorBridge emulatorBridge, Process process)
+        public GameQuitEventArgs(ICoreService eventCoreInstance, IGameInfo gameInfo, IEmulatorAssembly emulatorAssembly, IEmulatorBridge emulatorBridge)
             : base(eventCoreInstance, gameInfo)
         {
             this.GameEmulatorAssembly = emulatorAssembly;
             this.GameEmulatorBridge = GameEmulatorBridge;
-            this.GameEmulatorProcess = GameEmulatorProcess;
         }
