--- Snowflake.Events/CoreEvents/GameStartEventArgs.cs
@@ -1,25 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Snowflake.Service;
-using Snowflake.Emulator;
-using Snowflake.Game;
-
-namespace Snowflake.Events.CoreEvents
-{
-    public class GameStartEventArgs : GameEventArgs
-    {
-        public IEmulatorAssembly GameEmulatorAssembly { get; set; }
-        public IEmulatorBridge GameEmulatorBridge { get; set; }
-        public GameStartEventArgs(ICoreService eventCoreInstance, IGameInfo gameInfo, IEmulatorAssembly emulatorAssembly, IEmulatorBridge emulatorBridge)
-            : base(eventCoreInstance, gameInfo)
-        {
-            this.GameEmulatorAssembly = emulatorAssembly;
-            this.GameEmulatorBridge = GameEmulatorBridge;
-        }
-
-    
-    }
-}
